+++ CHANGELOG.md
@@ -2,6 +2,52 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.2.2](https://github.com/pahudnet/cdk-cloudfront-plus/compare/v0.2.1...v0.2.2) (2021-04-06)
+
+### [0.2.1](https://github.com/pahudnet/cdk-cloudfront-plus/compare/v0.1.35...v0.2.1) (2021-04-06)
+
+## [0.2.0](https://github.com/pahudnet/cdk-cloudfront-plus/compare/v0.1.35...v0.2.0) (2021-04-05)
+
+### [0.1.35](https://github.com/pahudnet/cdk-cloudfront-plus/compare/v0.1.34...v0.1.35) (2021-04-05)
+
+### [0.1.34](https://github.com/pahudnet/cdk-cloudfront-plus/compare/v0.1.33...v0.1.34) (2021-04-05)
+
+
+### Features
+
+* **ext): OAuth2 Authentication (Authorization Code Grant:** init ([#59](https://github.com/pahudnet/cdk-cloudfront-plus/issues/59)) ([6d5e440](https://github.com/pahudnet/cdk-cloudfront-plus/commit/6d5e44042dc864a138e943224db9e40c1fee7c19)), closes [#17](https://github.com/pahudnet/cdk-cloudfront-plus/issues/17)
+
+### [0.1.33](https://github.com/pahudnet/cdk-cloudfront-plus/compare/v0.1.32...v0.1.33) (2021-04-05)
+
+### [0.1.32](https://github.com/pahudnet/cdk-cloudfront-plus/compare/v0.1.31...v0.1.32) (2021-04-05)
+
+
+### Features
+
+* so8120 multiple origin ip retry ([#58](https://github.com/pahudnet/cdk-cloudfront-plus/issues/58)) ([ceb9bfc](https://github.com/pahudnet/cdk-cloudfront-plus/commit/ceb9bfcf8392d2964867f6c7d745154d180f140a))
+
+### [0.1.31](https://github.com/pahudnet/cdk-cloudfront-plus/compare/v0.1.30...v0.1.31) (2021-04-05)
+
+### [0.1.30](https://github.com/pahudnet/cdk-cloudfront-plus/compare/v0.1.29...v0.1.30) (2021-04-05)
+
+### [0.1.29](https://github.com/pahudnet/cdk-cloudfront-plus/compare/v0.1.28...v0.1.29) (2021-04-04)
+
+### [0.1.28](https://github.com/pahudnet/cdk-cloudfront-plus/compare/v0.1.27...v0.1.28) (2021-04-04)
+
+
+### Features
+
+* **SO8118:** Access Origin by Geolocation ([#52](https://github.com/pahudnet/cdk-cloudfront-plus/issues/52)) ([7eba0e9](https://github.com/pahudnet/cdk-cloudfront-plus/commit/7eba0e939a438b9a0066d4bea65ea00229f00849))
+
+### [0.1.27](https://github.com/pahudnet/cdk-cloudfront-plus/compare/v0.1.26...v0.1.27) (2021-04-04)
+
+### [0.1.26](https://github.com/pahudnet/cdk-cloudfront-plus/compare/v0.1.25...v0.1.26) (2021-04-03)
+
+
+### Features
+
+* **extension:** redirect by geolocation  ([#50](https://github.com/pahudnet/cdk-cloudfront-plus/issues/50)) ([012f591](https://github.com/pahudnet/cdk-cloudfront-plus/commit/012f59110be8f450e4b432f607cf49580b8e79f5)), closes [#11](https://github.com/pahudnet/cdk-cloudfront-plus/issues/11)
+
 ### [0.1.25](https://github.com/pahudnet/cdk-cloudfront-plus/compare/v0.1.24...v0.1.25) (2021-04-03)
 
 ### [0.1.24](https://github.com/pahudnet/cdk-cloudfront-plus/compare/v0.1.23...v0.1.24) (2021-04-03)